feat: organize utils
