+++ document-models/real-world-assets/real-world-assets.json
@@ -13,7 +13,7 @@
             "changeLog": [],
             "state": {
                 "global": {
-                    "schema": "type RealWorldAssetsState {\n    accounts: [Account!]!\n    principalLenderAccountId: ID! # Account.id (That relates to the Maker account)\n    spvs: [Spv!]!\n    serviceProviderFeeTypes: [ServiceProviderFeeType!]!\n    fixedIncomeTypes: [FixedIncomeType!]!\n    portfolio: [Asset!]!\n    transactions: [GroupTransaction!]!\n}\n\nunion Asset = FixedIncome | Cash\n\ntype FixedIncome {\n    id: ID!\n    fixedIncomeTypeId: ID! # FixedIncomeType.id\n    name: String!\n    spvId: ID! # Spv.id\n    maturity: DateTime!\n    purchaseDate: DateTime! # Weighted average of underlying transactions\n    notional: Float! # Face value sum of the underlying transactions\n    purchaseProceeds: Float! # Total cash outflow spent on purchases\n    salesProceeds: Float! #\n    purchasePrice: Float! # purchaseProceeds / notional\n    totalDiscount: Float! # notional - purchaseProceeds\n    realizedSurplus: Float! # Sum(Asset Proceeds) + Purchase Proceeds\n    ISIN: String\n    CUSIP: String\n    coupon: Float\n}\n\ntype Cash {\n    id: ID!\n    spvId: ID! # Spv.id\n    currency: String! # Will always be 'USD' for version 1\n    balance: Float! # Sum of all underlying transactions\n}\n\ntype FixedIncomeType {\n    id: ID!\n    name: String!\n}\n\ntype Spv {\n    id: ID!\n    name: String!\n}\n\ntype BaseTransaction {\n    id: ID!\n    assetId: ID! # Asset.id\n    amount: Float!\n    entryTime: DateTime!\n    tradeTime: DateTime\n    settlementTime: DateTime\n    txRef: String\n    accountId: ID # Account.id\n    counterPartyAccountId: ID # Account.id\n}\n\ntype Account {\n    id: ID!\n    reference: String! # e.g. bank account number or ETH address\n    label: String\n}\n\ntype TransactionFee {\n    id: ID!\n    serviceProviderFeeTypeId: ID!\n    amount: Float!\n}\n\ntype GroupTransaction {\n    id: ID!\n    type: GroupTransactionType!\n    entryTime: DateTime!\n    cashBalanceChange: Float!\n    fees: [TransactionFee!]\n    cashTransaction: BaseTransaction\n    fixedIncomeTransaction: BaseTransaction\n    feeTransactions: [BaseTransaction!]\n    interestTransaction: BaseTransaction\n}\n\nenum GroupTransactionType {\n    PrincipalDraw\n    PrincipalReturn\n    AssetPurchase\n    AssetSale\n    InterestPayment\n    FeesPayment\n}\n\ntype ServiceProviderFeeType {\n    id: ID!\n    name: String!\n    feeType: String!\n    accountId: ID! # Account.id\n}\n\nscalar DateTime",
+                    "schema": "type RealWorldAssetsState {\n    accounts: [Account!]!\n    principalLenderAccountId: ID! # Account.id (That relates to the Maker account)\n    spvs: [Spv!]!\n    serviceProviderFeeTypes: [ServiceProviderFeeType!]!\n    fixedIncomeTypes: [FixedIncomeType!]!\n    portfolio: [Asset!]!\n    transactions: [GroupTransaction!]!\n}\n\nunion Asset = FixedIncome | Cash\n\ntype FixedIncome {\n    id: ID!\n    fixedIncomeTypeId: ID! # FixedIncomeType.id\n    name: String!\n    spvId: ID! # Spv.id\n    maturity: DateTime!\n    purchaseDate: DateTime! # Weighted average of underlying transactions\n    notional: Float! # Face value sum of the underlying transactions\n    purchaseProceeds: Float! # Total cash outflow spent on purchases\n    salesProceeds: Float! #\n    purchasePrice: Float! # purchaseProceeds / notional\n    totalDiscount: Float! # notional - purchaseProceeds\n    realizedSurplus: Float! # Sum(Asset Proceeds) + Purchase Proceeds\n    ISIN: String\n    CUSIP: String\n    coupon: Float\n}\n\ntype Cash {\n    id: ID!\n    spvId: ID! # Spv.id\n    currency: String! # Will always be 'USD' for version 1\n    balance: Float! # Sum of all underlying transactions\n}\n\ntype FixedIncomeType {\n    id: ID!\n    name: String!\n}\n\ntype Spv {\n    id: ID!\n    name: String!\n}\n\ntype BaseTransaction {\n    id: ID!\n    assetId: ID! # Asset.id\n    amount: Float!\n    entryTime: DateTime!\n    tradeTime: DateTime\n    settlementTime: DateTime\n    txRef: String\n    accountId: ID # Account.id\n    counterPartyAccountId: ID # Account.id\n}\n\ntype Account {\n    id: ID!\n    reference: String! # e.g. bank account number or ETH address\n    label: String\n}\n\ntype TransactionFee {\n    id: ID!\n    serviceProviderFeeTypeId: ID!\n    amount: Float!\n}\n\ntype GroupTransaction {\n    id: ID!\n    type: GroupTransactionType!\n    entryTime: DateTime!\n    cashBalanceChange: Float!\n    fees: [TransactionFee!]\n    cashTransaction: BaseTransaction\n    fixedIncomeTransaction: BaseTransaction\n}\n\nenum GroupTransactionType {\n    PrincipalDraw\n    PrincipalReturn\n    AssetPurchase\n    AssetSale\n    InterestPayment\n    FeesPayment\n}\n\ntype ServiceProviderFeeType {\n    id: ID!\n    name: String!\n    feeType: String!\n    accountId: ID! # Account.id\n}\n\nscalar DateTime",
                     "initialValue": "{\"accounts\":[{\"id\":\"maker-principal-account-id\",\"reference\":\"Maker Principal\",\"label\":\"Maker Principal\"}],\"principalLenderAccountId\":\"maker-principal-account-id\",\"spvs\":[],\"serviceProviderFeeTypes\":[],\"fixedIncomeTypes\":[],\"portfolio\":[{\"id\":\"maker-principal-asset-id\",\"spvId\":\"1\",\"currency\":\"USD\",\"balance\":0}],\"transactions\":[]}",
                     "examples": []
                 },
